HWKBTM-24 Refactoring to support deployment to standalone app and wildfly environment
