+++ api/src/test/java/org/hawkular/btm/api/util/ServiceResolverTest.java
@@ -33,8 +33,8 @@ public class ServiceResolverTest {
 
     @Test
     public void testSingletonService() {
-        TestService s1 = ServiceResolver.getSingletonService(TestService.class);
-        TestService s2 = ServiceResolver.getSingletonService(TestService.class);
+        TestService s1 = ServiceResolver.getSingletonService(TestService.class).getNow(null);
+        TestService s2 = ServiceResolver.getSingletonService(TestService.class).getNow(null);
 
         assertNotNull(s1);
         assertNotNull(s2);
@@ -51,9 +51,8 @@ public class ServiceResolverTest {
 
     @Test
     public void testNoImplementation() {
-        Map<?,?> s1 = ServiceResolver.getSingletonService(Map.class);
+        Map<?,?> s1 = ServiceResolver.getSingletonService(Map.class).getNow(null);
 
         assertNull(s1);
     }
-
 }
