--- client/btxn-client/src/main/java/org/hawkular/btm/btxn/client/config/InstrumentTypeTransformer.java
@@ -1,46 +1,0 @@
-/*
- * Copyright 2015 Red Hat, Inc. and/or its affiliates
- * and other contributors as indicated by the @author tags.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *    http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.hawkular.btm.btxn.client.config;
-
-import org.hawkular.btm.api.model.admin.InstrumentType;
-
-/**
- * This interface represents a transformer to convert the instrumentation
- * type into a ByteMan rule.
- *
- * @author gbrown
- */
-public interface InstrumentTypeTransformer {
-
-    /**
-     * This method returns the instrument type associated with the
-     * transformer.
-     *
-     * @return The instrument type
-     */
-    Class<? extends InstrumentType> getInstrumentType();
-
-    /**
-     * This method converts the supplied instrumentation type into a
-     * ByteMan rule.
-     *
-     * @param type The instrument type
-     * @return The ByteMan rule
-     */
-    String convertToRule(InstrumentType type);
-
-}
