--- client/btxn-client/src/main/resources/META-INF/services/org.hawkular.btm.api.client.ConfigurationManager
@@ -1,1 +1,0 @@
-org.hawkular.btm.btxn.client.internal.StandaloneConfigurationManagerImpl
