--- client/btxn-client/src/main/resources/META-INF/services/org.hawkular.btm.btxn.client.config.InstrumentTypeTransformer
@@ -1,1 +1,0 @@
-org.hawkular.btm.btxn.client.config.InstrumentMethodTransformer
