+++ client/btxn-collector/src/main/resources/META-INF/beans.xml
