+++ client/btxn-collector/src/main/resources/META-INF/services/org.hawkular.btm.api.client.BusinessTransactionCollector
@@ -1,0 +1,1 @@
+org.hawkular.btm.client.collector.DefaultBusinessTransactionCollector
