+++ client/client-manager/src/main/resources/META-INF/services/org.hawkular.btm.client.manager.config.InstrumentTypeTransformer
@@ -1,0 +1,1 @@
+org.hawkular.btm.client.manager.config.InstrumentMethodTransformer
