+++ client/client-se-rest/src/main/scripts/all-assembly.xml
