+++ client/config-manager-rest-client/src/main/resources/META-INF/services/org.hawkular.btm.api.client.ConfigurationManager
@@ -1,0 +1,1 @@
+org.hawkular.btm.client.config.rest.client.ConfigurationManagerRESTClient
