+++ tests/wildfly/pom.xml
@@ -70,7 +70,7 @@
     </dependency>
     <dependency>
       <groupId>org.hawkular.btm</groupId>
-      <artifactId>hawkular-btm-btxn-client</artifactId>
+      <artifactId>hawkular-btm-client-manager</artifactId>
       <scope>test</scope>
     </dependency>
 
@@ -185,6 +185,8 @@
                   <outputDirectory>${hawkular.deployments}</outputDirectory>
                   <destFileName>camel-example-servlet-rest-tomcat.war</destFileName>
                 </artifactItem>
+
+                <!-- BTM client configuration -->
                 <artifactItem>
                   <groupId>org.jboss.byteman</groupId>
                   <artifactId>byteman</artifactId>
@@ -193,39 +195,11 @@
                   <destFileName>byteman.jar</destFileName>
                 </artifactItem>
                 <artifactItem>
-                  <groupId>com.fasterxml.jackson.core</groupId>
-                  <artifactId>jackson-core</artifactId>
-                  <overWrite>true</overWrite>
-                  <outputDirectory>${hawkular.lib}</outputDirectory>
-                  <destFileName>jackson-core.jar</destFileName>
-                </artifactItem>
-                <artifactItem>
-                  <groupId>com.fasterxml.jackson.core</groupId>
-                  <artifactId>jackson-databind</artifactId>
-                  <overWrite>true</overWrite>
-                  <outputDirectory>${hawkular.lib}</outputDirectory>
-                  <destFileName>jackson-databind.jar</destFileName>
-                </artifactItem>
-                <artifactItem>
-                  <groupId>com.fasterxml.jackson.core</groupId>
-                  <artifactId>jackson-annotations</artifactId>
-                  <overWrite>true</overWrite>
-                  <outputDirectory>${hawkular.lib}</outputDirectory>
-                  <destFileName>jackson-annotations.jar</destFileName>
-                </artifactItem>
-                <artifactItem>
                   <groupId>org.hawkular.btm</groupId>
-                  <artifactId>hawkular-btm-api</artifactId>
+                  <artifactId>hawkular-btm-client-se-rest</artifactId>
                   <overWrite>true</overWrite>
                   <outputDirectory>${hawkular.lib}</outputDirectory>
-                  <destFileName>hawkular-btm-api.jar</destFileName>
-                </artifactItem>
-                <artifactItem>
-                  <groupId>org.hawkular.btm</groupId>
-                  <artifactId>hawkular-btm-btxn-client</artifactId>
-                  <overWrite>true</overWrite>
-                  <outputDirectory>${hawkular.lib}</outputDirectory>
-                  <destFileName>hawkular-btm-btxn-client.jar</destFileName>
+                  <destFileName>hawkular-btm-client-se-rest.jar</destFileName>
                 </artifactItem>
               </artifactItems>
             </configuration>
@@ -333,7 +307,7 @@
                 <javaOpt>-Xrunjdwp:transport=dt_socket,server=y,suspend=n,address=8787</javaOpt>
                 <javaOpt>-Dkeycloak.import=${hawkular.configuration}/hawkular-realm.json</javaOpt>
                 <javaOpt>-Dorg.jboss.byteman.transform.all</javaOpt>
-                <javaOpt>-javaagent:${hawkular.lib}/byteman.jar=manager:org.hawkular.btm.btxn.client.ClientManager,boot:${hawkular.lib}/byteman.jar,sys:${hawkular.lib}/jackson-core.jar,sys:${hawkular.lib}/jackson-annotations.jar,sys:${hawkular.lib}/jackson-databind.jar,sys:${hawkular.lib}/hawkular-btm-api.jar,sys:${hawkular.lib}/hawkular-btm-btxn-client.jar</javaOpt>
+                <javaOpt>-javaagent:${hawkular.lib}/byteman.jar=manager:org.hawkular.btm.client.manager.ClientManager,boot:${hawkular.lib}/byteman.jar,sys:${hawkular.lib}/hawkular-btm-client-se-rest.jar</javaOpt>
                 <javaOpt>-Dhawkular-btm.config=${hawkular.configuration}/hawkular-btm-config.json</javaOpt>
               </javaOpts>
               <startup-timeout>120</startup-timeout>
